### 2021-01-20 #2944 @kthyng
Hi @LangeJens! We are wrapping up our scope query for your submission. One suggestion that came up a couple of times is whether this software could be integrated into a larger R package? Is there one that would be a logical choice for this? The idea is that smaller software would be used much more often potentially in a larger package that can cover more functions.
