### 2021-01-20 #2944 @LangeJens
Hi @kthyng! Thank you very much for getting in touch. I am very thankful that you take the time to carefully consider my submission. Regarding your question: I did not yet program another R package of a broader scope into which CliquePercolation could be intergrated. Moreover, I am not collarborating with someone who developed a package into which CliquePercolation could be merged. Even though I wish to extend the package in the future with another popular method for identifying mixed communities, I cannot predict when (or even whether) this is going to happen.
Nevertheless, I want to emphasize that the clique percolation algorithm is popular and frequently used in practice. As a rough indication, the original paper introducing the classic algorithm was cited more than 5,000 times. Furthermore, the CliquePercolation package is downloaded 400 to 500 times each month, speaking to its popularity despite the fact that it is - admittedly - specialized and small compared to other R packages.
