### 2021-01-20 #2944 @kthyng
@LangeJens Ok, thanks for your thorough response. Let us proceed with the review then. We see that the package is on the small side for JOSS, but we have also seen that it appears to fill a software need.
