### 2021-02-04 #2944 @GiulioCostantini
Hi all,
Unfortunately the same the same is true for me. I would suggest also Jonas
Haslbeck (jonashaslbeck@gmail.com) as another alternative reviewer,
All the best
Giulio

Giulio Costantini, PhD
Assistant professor
Psychology Department
University of Milano-Bicocca
P.zza dell'Ateneo Nuovo, 1 (U6) - room 3127
20126 Milan (MI)
Phone ++39 02.6448.3839

<https://www.avast.com/sig-email?utm_medium=email&utm_source=link&utm_campaign=sig-email&utm_content=webmail>
Mail
priva di virus. www.avast.com
<https://www.avast.com/sig-email?utm_medium=email&utm_source=link&utm_campaign=sig-email&utm_content=webmail>
<#DAB4FAD8-2DD7-40BB-A1B8-4E2AA1F9FDF2>

Il giorno mer 3 feb 2021 alle ore 22:13 Payton J. Jones <
notifications@github.com> ha scritto:

> Unfortunately, I don't think I would be able to review this in a reasonable
> time frame. Sacha Epskamp (sacha.epskamp@gmail.com) or Donald Williams (
> drwwilliams@ucdavis.edu) might be good alternate reviewers.
>
> On Tue, Feb 2, 2021 at 5:47 PM Karthik Ram <notifications@github.com>
> wrote:
>
> > 👋 @paytonjjones <https://github.com/paytonjjones>
> > Would you be able to review this submission?
> >
> > —
> > You are receiving this because you were mentioned.
> > Reply to this email directly, view it on GitHub
> > <
> https://github.com/openjournals/joss-reviews/issues/2944#issuecomment-772066570
> >,
> > or unsubscribe
> > <
> https://github.com/notifications/unsubscribe-auth/AFQWBJPIQ7K7MMGIMSVMW6DS5B6JTANCNFSM4VXXNYKQ
> >
> > .
> >
>
