### 2021-02-04 #2944 @LangeJens
I want to mention, that I know Sacha Epskamp and Jonas Haslbeck from my time in Amsterdam. I worked in another department at the same university and we met in group as well as personal meetings. I do not necessarily believe that they would have a conflict of interest, but I wanted to disclose this in case it is relevant.
