### 2021-03-01 #2944 @donaldRwilliams
@karthik I noted the software has under 1,000 lines of R code, yet it was not flagged according to the guidelines, in addition to having 25 commits.

I am happy to review anyhow, but want to be sure it is worth my time given the length of code and submit history. Thanks !
