### 2021-03-19 #2944 @LangeJens
Thanks for getting in touch! If it is of any help, you could also ask Julian Burger or Riet Van Bork from the University of Amsterdam. I know that they also contribute to R packages although they do not seem to have github pages themselves.
