### 2021-04-22 #3210 @whedon
[REVIEW]: CliquePercolation: An R Package for conducting and visualizing results of the clique percolation network community detection algorithm
**Submitting author:** @LangeJens (<a href="http://orcid.org/0000-0002-5375-3247">Jens Lange</a>)
**Repository:** <a href="https://github.com/LangeJens/CliquePercolation" target ="_blank">https://github.com/LangeJens/CliquePercolation</a>
**Version:** v0.3.0
