### 2021-05-03 0ae5ab3
JOSS Code Review
## Alex changes for JOSS review

o For `cpAlgorithm`, `cpPermuteEntropy`, and `cpThreshold`, I made it so that any matrix can be input (rather than just qgraph objects). The code still does not specify for a square or symmetric matrix, which would be recommended. Having input as either a matrix (data.frame) or qgraph object allows your users greater flexibility

o Specific for `cpPermuteEntropy`: I added code to parallelize the function to increase the computational speed. Although the old code worked well for a small number of variables, I performed the analysis on one of my 
[…]
