+++ R/cpPermuteEntropy.R
@@ -11,6 +11,11 @@
 #'    default is 0.95)
 #' @param CFinder logical indicating whether clique percolation for weighted networks
 #'    should be performed as in CFinder ; see also \link{cpAlgorithm}
+#' @param ncores Numeric.
+#' Number of cores to use in computing results.
+#' Defaults to \code{parallel::detectCores() / 2} or half of your
+#' computer's processing power.
+#' Set to \code{1} to not use parallel computing
 #'    
 #' @return
 #'   A list object with the following elements:
@@ -73,11 +78,23 @@
 #' @export cpPermuteEntropy
 
 cpPermuteEntropy <- function(W, cpThreshold.object, n = 100, interval = 0.95,
-                             CFinder = FALSE) {
+                             CFinder = FALSE, ncores) {
   
   #error message if W is not a qgraph object
-  if (methods::is(W, "qgraph") == FALSE) {
-    stop("W (network object) must be a qgraph object.")
+  if (!isTRUE(methods::is(W, "qgraph"))) {
+    
+    #### ALEX JOSS REVIEW CODE BEGIN ####
+    #converts matrix to qgraph if input is not a qgraph object
+    W <- qgraph::qgraph(W, DoNotPlot = TRUE)
+    
+    #### ALEX JOSS REVIEW CODE END ####
+    
+    #### OLD ERROR BEGIN ####
+    
+    #stop("W (network object) must be a qgraph object.")
+    
+    #### OLD ERROR END ####
+    
   }
   
   #error message if n is not larger than zero
@@ -89,6 +106,15 @@ cpPermuteEntropy <- function(W, cpThreshold.object, n = 100, interval = 0.95,
   if (!(interval < 1 & interval > 0)) {
     stop("interval must be larger than zero and smaller than 1.")
   }
+  
+  #### ALEX JOSS REVIEW CODE BEGIN ####
+  
+  #set default for missing processing cores
+  if(missing(ncores)){
+    ncores <- ceiling(parallel::detectCores() / 2)
+  }
+  
+  #### ALEX JOSS REVIEW CODE END ####
   
   #for n random permutations of graph, the function runs cpTthreshold for k.range and I.range
   #random permutations generated by permuting the upper triangle of the weights matrix and forcing it symmetric
@@ -102,38 +128,97 @@ cpPermuteEntropy <- function(W, cpThreshold.object, n = 100, interval = 0.95,
   }
   k.range <- sort(unique(cpThreshold.object$k))
   I.range <- sort(unique(cpThreshold.object$Intensity), decreasing = TRUE)
-  max_ent <- data.frame(k = numeric(),
-                        max = numeric())
-  counter <- 1
   
-  progress_bar_permute <- utils::txtProgressBar(min = 0, max = 100,
-                                                style = 3) #progress bar definition
-  progress_bar_permute_counter <- 0 #counter for progress bar
   
-  for (i in 1:n) {
-    Wmat <- qgraph::getWmat(W)
-    Wmat[upper.tri(Wmat)] <- sample(Wmat[upper.tri(Wmat)])
-    Wmat <- Matrix::forceSymmetric(Wmat, "U")
-    graph <- qgraph::qgraph(Wmat, DoNotPlot = TRUE)
+  #### ALEX JOSS REVIEW CODE BEGIN ####
+  
+  #initialize iterations
+  iterations <- as.list(1:n)
+  
+  #parallel processing
+  cl <- parallel::makeCluster(ncores)
+  
+  #export variables
+  parallel::clusterExport(cl = cl,
+                          varlist = c("W", "k.range", "I.range",
+                                      "method", "cpThreshold"),
+                          envir=environment())
+  
[…]
