### 2021-05-03 #3210 @AlexChristensen
@LangeJens and @karthik,

I've gone through the review checklist and see that everything is OK.

I ran through the example in the manuscript with some of my own data and made a few updates to the packages code. Here are the changes that I made in my fork: https://github.com/AlexChristensen/CliquePercolation

## Alex changes for JOSS review

o For `cpAlgorithm`, `cpPermuteEntropy`, and `cpThreshold`, I made it so that any matrix can be input (rather than just *qgraph* objects). The code still does not specify for a square or symmetric matrix, which would be recommended. Having input as either a matrix (data.frame) or *qgraph* object allows your users greater flexibility

o Specific for `cpPermuteEntropy`: I added code to parallelize the function to increase the computational speed. Although the old code worked well for a small number of variables, I performed the analysis on one of my own datasets with 48 variables and this took 82.16 seconds. With the parallelization, the same analysis took 22.25 seconds (or about four times faster). The parallelization should scale with the sample size. Here's my computer specs: Windows 10 (x64), Intel i7-9750H 2.60GHz, 32GB RAM

o Note that I changed the DESCRIPTION file to add the *parallel* package as an IMPORT for the parallelization

o I also changed the .Rd of `cpPermuteEntropy` to add the `ncores` argument. The default is for half the processing cores on the users computer

@LangeJens, I can create a pull request if these changes are acceptable. Feel free to change the argument name and description for the `ncores` argument as necessary. I've commented out the original code to retain previous versioning. I've also tested the changes on a number of datasets to make sure that my additions do not introduce any errors.
