### 2021-05-04 #3210 @LangeJens
@AlexChristensen and @karthik,

Thank you very much Alex. I am very grateful that you took the time to improve my code. Especially the implementation of parallel computing is amazing! I would be happy if you could create a pull request that I can then merge with the master branch. After merging, I will add a few other changes:

- Mention in the .Rd of cpPermuteEntropy, cpAlgorithm, and cpThreshold that users can also specify a matrix for W.
- Include checks in the above functions whether the matrix is (a) square (comparing number of rows with number of columns) and (b) symmetric (using isSymmetric()).
- Include pbapply in DESCRIPTION IMPORT as Alex used it during parallel computing.
- Remove Alex' comments in the code.
- Add Alex as a contributor in DESCRIPTION.

In case you disagree with any of the changes, please let me know.
