### 2021-05-06 ab5f559
Extend JOSS Review Changes
- Mention that users can specify matrix
- Include checks for matrix
- Adapt DESCRIPTION
