+++ R/cpPermuteEntropy.R
@@ -4,7 +4,7 @@
 #' community partition from clique percolation based on randomly permuted networks
 #' of original network.
 #' 
-#' @param W A qgraph object; see also \link[qgraph]{qgraph}
+#' @param W A qgraph object or a symmetric square matrix; see also \link[qgraph]{qgraph}
 #' @param cpThreshold.object A cpThreshold object; see also \link{cpThreshold}
 #' @param n number of permutations (default is 100)
 #' @param interval requested confidence interval (larger than zero and smaller 1;
@@ -12,10 +12,10 @@
 #' @param CFinder logical indicating whether clique percolation for weighted networks
 #'    should be performed as in CFinder ; see also \link{cpAlgorithm}
 #' @param ncores Numeric.
-#' Number of cores to use in computing results.
-#' Defaults to \code{parallel::detectCores() / 2} or half of your
-#' computer's processing power.
-#' Set to \code{1} to not use parallel computing
+#'    Number of cores to use in computing results.
+#'    Defaults to \code{parallel::detectCores() / 2} or half of your
+#'    computer's processing power.
+#'    Set to \code{1} to not use parallel computing
 #'    
 #' @return
 #'   A list object with the following elements:
@@ -80,14 +80,22 @@
 cpPermuteEntropy <- function(W, cpThreshold.object, n = 100, interval = 0.95,
                              CFinder = FALSE, ncores) {
   
-  #error message if W is not a qgraph object
+  ###check whether W is a qgraph object
+  ###if not check whether matrix is symmetric and square and convert to qgraph object
   if (!isTRUE(methods::is(W, "qgraph"))) {
     
-    #### ALEX JOSS REVIEW CODE BEGIN ####
+    #error if W is a matrix but not symmetric
+    if (isSymmetric(W) == FALSE) {
+      stop("If W is a matrix, it must be symmetric.")
+    }
+    
+    #error if W is a matrix but not square
+    if (nrow(W) != ncol(W)) {
+      stop("If W is a matrix, it must be a square matrix.")
+    }
+    
     #converts matrix to qgraph if input is not a qgraph object
     W <- qgraph::qgraph(W, DoNotPlot = TRUE)
-    
-    #### ALEX JOSS REVIEW CODE END ####
     
     #### OLD ERROR BEGIN ####
     
@@ -107,16 +115,12 @@ cpPermuteEntropy <- function(W, cpThreshold.object, n = 100, interval = 0.95,
     stop("interval must be larger than zero and smaller than 1.")
   }
   
-  #### ALEX JOSS REVIEW CODE BEGIN ####
-  
   #set default for missing processing cores
   if(missing(ncores)){
     ncores <- ceiling(parallel::detectCores() / 2)
   }
   
-  #### ALEX JOSS REVIEW CODE END ####
-  
-  #for n random permutations of graph, the function runs cpTthreshold for k.range and I.range
+  #for n random permutations of graph, the function runs cpThreshold for k.range and I.range
   #random permutations generated by permuting the upper triangle of the weights matrix and forcing it symmetric
   #for each random permutation, the maximum entropy is extracted for each k and saved in data frame
   #information about parameters in original cpThreshold function is extracted from cpThreshold.object
@@ -128,9 +132,6 @@ cpPermuteEntropy <- function(W, cpThreshold.object, n = 100, interval = 0.95,
   }
   k.range <- sort(unique(cpThreshold.object$k))
   I.range <- sort(unique(cpThreshold.object$Intensity), decreasing = TRUE)
-  
-  
-  #### ALEX JOSS REVIEW CODE BEGIN ####
   
[…]
