+++ paper/paper.bib
@@ -69,6 +69,10 @@
   title = {Emotions as Overlapping Causal Networks of Emotion Components: Implications and Methodological Approaches},
   journal = {Emotion Review},
   year = {2021},
+  volume = {13},
+  number = {2},
+  pages = {157--167},
+  doi = {10.1177/1754073920988787},
 }
 
 @Article{Palla+et+al:2005,
