### 2021-05-06 #3210 @LangeJens
I merged the pull request, implemented all the changes I listed above, and pushed the changes to the master branch. I also updated the reference in the paper that did not yet have a doi assigned when I submitted.

@AlexChristensen I noticed during testing that setting a seed with set.seed() during parallel computing does not work. I could not find an easy fix online. Do you have a recommendation? Your help would be highly appreciated.
