### 2021-05-11 c93e4a0
Delete unnessecary code
