### 2021-05-11 6d67bc3
JOSS Review
o added code and documentation to set seed for reproducible paralellization
