### 2021-05-11 #3210 @AlexChristensen
@LangeJens ,

I've added code that will allow for reproducible parallelization (see latest pull request). This was done using the `clusterSetRNGSream` function in the *parallel* package.

With this, I added an argument called `seed` that defaults to `NULL`. The default will produce stochastic results. To get reproducible results, you can set the `seed` argument.

Like the example in the paper, you can use:

```
thresholds.permute <- cpPermutateEntropy(W, cpThreshold.object = threshold, ncores = 2, seed = 4186)
```

You would no longer need to use `set.seed` unless you were wanting to set the seed for the rest of the functions.
