### 2021-05-15 LangeJens/CliquePercolation#3 @donaldRwilliams
real world examples 
I think the examples can be improved to including "real world" examples. Or perhaps one actual example with real data. As it is, constructing the `W` seems a bit unsatisfying
