### 2021-05-15 #3210 @donaldRwilliams
@karthik I have opened up 8 issues on the repo. I think the proposed changes will improve the package considerably.

1.  Currently, the package contains no real world examples. (https://github.com/LangeJens/CliquePercolation/issues/3). Rather the examples are "fake" networks when it would be ideal to have at least one example with real data, I think.
