### 2021-05-15 #3210 @donaldRwilliams
@karthik @LangeJens Note the above suggestions on the one hand are "cosmetic", but on the other, I find that nicely developed software typically include all the above. I am confident that making those changes will greatly improve the users experience and potentially increase impact (e.g., by honing in the README).
