### 2021-05-16 eb97b2a
Add continuous integration and badges
