+++ .travis.yml
@@ -1,5 +1,5 @@
-# R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
-
 language: R
+
 cache: packages
+
 warnings_are_errors: false
