### 2021-05-17 5e45127
Adapt help files to possible matrix specification
