+++ man/Obama.Rd
@@ -1,0 +1,44 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/Obama.R
+\docType{data}
+\name{Obama}
+\alias{Obama}
+\title{Data: Evaluative Reactions Toward Barack Obama (2012)}
+\format{
+An object of class \code{"data.frame"} with 5,914 observations
+   and 10 variables.
+   \describe{
+     \item{Mor}{"Is moral"}
+     \item{Led}{"Would provide strong leadership"}
+     \item{Car}{"Really cares about people like you"}
+     \item{Kno}{"Is knowledgeable"}
+     \item{Int}{"Is intelligent"}
+     \item{Hns}{"Is honest"}
+     \item{Ang}{"Angry"}
+     \item{Hop}{"Hopeful"}
+     \item{Afr}{"Afraid of him"}
+     \item{Prd}{"Proud"}
+   }
+}
+\source{
+\url{www.electionstudies.org}
+}
+\usage{
+data(Obama)
+}
+\description{
+A data set containing evaluative reactions toward
+Barack Obama from the American National Election Studies
+in 2012. The study included 5,914 participants, representative
+of the adult U.S. American population (note missing values).
+The participants rated Obama on 10 evaluative reactions.
+}
+\examples{
+data(Obama)
+}
+\references{
+Dalege, J., Borsboom, D., Van Harreveld, F., Van der Maas, H. L. J. (2017).
+Network analysis on attitudes: A brief tutorial. \emph{Social Psychological
+and Personality Science, 8}, 528-537. https://doi.org/10.1177/1948550617709827
+}
+\keyword{datasets}
