+++ man/immuno.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/immuno.R
+\docType{data}
+\name{immuno}
+\alias{immuno}
+\title{Data: Immunoglobulin interaction network}
+\format{
+An object of class \code{"qgraph"} with 1,316 nodes and
+   6,300 edges.
+}
+\source{
+\url{https://CRAN.R-project.org/package=igraphdata}
+}
+\usage{
+data(immuno)
+}
+\description{
+Unweighted, undirected network of interactions in the immunoglobulin
+network. The 1,316 nodes represent amino-acids and two nodes are
+connected by an edge if the shortest distance of their C_alpha atoms
+is smaller than \eqn{\Theta = 8} Angstrom.
+}
+\examples{
+data(immuno)
+}
+\references{
+Gfeller, D. (2007). \emph{Simplifying complex networks: From a clustering
+to a coarse graining strategy}. EPFL. http://library.epfl.ch/theses/?nr=3888
+}
+\keyword{datasets}
