+++ DESCRIPTION
@@ -14,6 +14,7 @@ Authors@R: c(
 Maintainer: Jens Lange <lange.jens@outlook.com>
 License: GPL-3
 Encoding: UTF-8
+LazyData: true
 Depends:
    R (>= 3.6.0)
 Imports:
