### 2021-05-18 5968bb6
Return objects as invisible
