+++ R/cpCommunitySizeDistribution.R
@@ -25,7 +25,7 @@
 #'   
 #'   If \code{test.power.law = TRUE}, test of a fit of a power-law is performed with the
 #'   function fit_power_law (see also \link[igraph]{fit_power_law}). Fit is tested for
-#'   the entire distribution from the smallest community size onwards (i.e., typically k
+#'   the entire distribution from the smallest community size onward (i.e., typically k
 #'   as specified in cpAlgorithm). Moreover, test uses the \code{plfit} implementation of
 #'   fit_power_law. For other arguments, default values are used.
 #' 
@@ -84,15 +84,13 @@ cpCommunitySizeDistribution <- function(list.of.communities, color.line = "#bc00
     graphics::lines(x = size_freq$size, y = size_freq$frequency, col = color.line, lwd = 3)
     
     if (test.power.law == FALSE) {
-      return(list(size.distribution = size_freq))
-    }
-    
-    if (test.power.law == TRUE) {
+      invisible(list(size.distribution = size_freq))
+    } else {
       size_vector <- rep(size_freq$size, size_freq$frequency)
       fit_pl <- igraph::fit_power_law(size_vector, xmin = min(size_vector),
                                       implementation = "plfit")
-      return(list(size.distribution = size_freq,
-                  fit.power.law = fit_pl))
+      invisible(list(size.distribution = size_freq,
+                     fit.power.law = fit_pl))
     }
   }
   
