+++ NAMESPACE
@@ -1,5 +1,8 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(print,cpAlgorithm)
+S3method(print,cpPermuteEntropy)
+S3method(summary,cpAlgorithm)
 export(cpAlgorithm)
 export(cpColoredGraph)
 export(cpCommunityGraph)

[…]
