### 2021-05-18 5c45a16
Add real-world examples to all functions
