+++ man/cpAlgorithm.Rd
@@ -78,6 +78,12 @@ W <- qgraph::qgraph(W)
 # run clique percolation for unweighted networks
 results <- cpAlgorithm(W = W, k = 3, method = "unweighted")
 
+# print results overview
+results
+
+# extract more information about communities
+summary(results)
+
 ## Example for weighted networks
 
 # create qgraph object
@@ -97,6 +103,29 @@ W <- qgraph::qgraph(W)
 
 # run clique percolation for weighted networks
 results <- cpAlgorithm(W = W, k = 3, method = "weighted", I = 0.1)
+
+# print results overview
+results
+
+# extract more information about communities
+summary(results)
+
+## Example with Obama data set (see ?Obama)
+
+# get data
+data(Obama)
+
+# estimate network
+net <- qgraph::EBICglasso(qgraph::cor_auto(Obama), n = nrow(Obama))
+
+# run clique percolation algorithm with specific k and I
+cpk3I.16 <- cpAlgorithm(net, k = 3, I = 0.16, method = "weighted")
+
+# print results overview
+cpk3I.16
+
+# extract more information about communities
+summary(cpk3I.16)
 
 }
 \references{
