### 2021-05-18 LangeJens/CliquePercolation#3 @LangeJens
I added a real-world example with one of the included data sets in each function.
