### 2021-05-19 450bb62
Include print and summary functions in vignette
