+++ vignettes/CliquePercolation.Rmd
@@ -216,7 +216,7 @@ In sum, the clique percolation algorithm proceeds by identifying $k$-cliques and
 
 The **CliquePercolation** package entails functions to conduct the clique percolation community detection algorithms for unweighted and weighted networks and to interpret the results. In what follows, an example is used to explain the workflow suggested by the package.
 
-**CliquePercolation** accepts only networks that were analyzed with the **qgraph** package [@Epskamp+et+al:2012]. **qgraph** accepts output from various different packages such as **igraph** [@Csardi+Nepusz:2006], **bootnet** [@Epskamp+Fried:2018], or **lavaan** [@Rosseel:2012]. For details see the documentation of **qgraph**. Moreover, I will use the **Matrix** [@Bates+Maechler:2019] package for the example below. Thus, we need to load the following packages in **R** to run the example
+**CliquePercolation** accepts networks that were analyzed with the **qgraph** package [@Epskamp+et+al:2012]. **qgraph** accepts output from various different packages such as **igraph** [@Csardi+Nepusz:2006], **bootnet** [@Epskamp+Fried:2018], or **lavaan** [@Rosseel:2012]. For details see the documentation of **qgraph**. Moreover, I will use the **Matrix** [@Bates+Maechler:2019] package for the example below. Thus, we need to load the following packages in **R** to run the example
 
 ```{r setup}
 library(CliquePercolation) #version 0.3.0
@@ -224,7 +224,7 @@ library(qgraph)            #version 1.6.5
 library(Matrix)            #version 1.2-18
 ```
 
-First, I will create an example network. When you want to apply the clique percolation algorithm, you will already have a network. Importantly, it needs to be a **qgraph** object. The network I am using is weighted, with edge weights drawn from a random distribution with mean of 0.3 and a standard deviation of 0.1.
+First, I will create an example network. When you want to apply the clique percolation algorithm, you will already have a network. The network can be a **qgraph** object. Alternatively, you can also provide a symmetric matrix with rows and columns representing the nodes and each entry representing the edge connecting the respective nodes (i.e., an adjacency matrix or a weights matrix). The network I am using is weighted, with edge weights drawn from a random distribution with mean of 0.3 and a standard deviation of 0.1.
 
 ```{r, echo = TRUE, dpi = 400, fig.cap = "**Weighted network with 150 nodes.**", fig.align = "center", out.extra = 'style = "border:none"', out.width = "100%"}
 # create qgraph object; 150 nodes with letters as names; 1/7 of edges different from zero
@@ -246,7 +246,7 @@ W <- qgraph::qgraph(W, theme = "colorblind", layout = "spring", cut = 0.4)
 
 &nbsp;
 
-To run the clique percolation algorithm for weighted networks, we initially need to optimize $k$ and $I$. To this end, the `cpThreshold` function can be used. For a specific network (i.e., a **qgraph** object), it determines the ratio of the largest to second largest community, $\chi$, and/or entropy over a range of $k$ and $I$ values. It does so by running the clique percolation algorithm for either unweighted (`method = "unweighted"`) or weighted networks (`method = "weighted"`). The algorithm as implemented in **CFinder** can also be requested (`method = "weighted.CFinder"`). For each $k$ and $I$ combination, `cpThreshold` determines the respective threshold and saves the results in a data frame, specifying the respective $k$, $I$, the number of communities, the number of isolated nodes, and the requested thresholds. Note that the ratio threshold can be determined only when there
[…]
