+++ README.Rmd
@@ -1,0 +1,158 @@
+---
+output: github_document
+---
+
+```{r, echo = FALSE, message = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>",
+  dev = "png",
+  dpi = 200,
+  fig.align = "center",
+  knitr::opts_chunk$set(comment = NA)
+  )
+library(qgraph)
+library(bootnet)
+library(CliquePercolation)
+```
+
+[![CRAN Version](http://www.r-pkg.org/badges/version/CliquePercolation)](https://cran.r-project.org/package=CliquePercolation)
+[![Downloads](https://cranlogs.r-pkg.org/badges/CliquePercolation)](https://cran.r-project.org/package=CliquePercolation)
+[![Travis build status](https://travis-ci.com/LangeJens/CliquePercolation.svg?branch=master)](https://travis-ci.com/LangeJens/CliquePercolation)
+
+# CliquePercolation
+
+**CliquePercolation** is an **R** package that entails multiple functions related to the clique percolation community detection algorithms for undirected, unweighted networks and for undirected, weighted networks. As various sciences analyze structural properties of networks, the **CliquePercolation** package can be useful in many fields such as physics, biology, psychology, computer science, or sociology.
+
+## Installation
+
+To use **CliquePercolation**, you first need to install **R** (find download page [here](https://cran.r-project.org/)). Moreover, I recommend to use **R** via *RStudio*, an environment that facilitates the use of **R** in multiple ways (find download page [here](https://rstudio.com/)).
+
+The developmental version of **CliquePercolation** is available on GitHub. You can download it from GitHub directly with the help of the **devtools** package in **R**:
+
+
+```{r, eval = FALSE}
+install.packages("devtools")
+devtools::install_github("LangeJens/CliquePercolation", build_vignettes = TRUE)
+```
+
+
+You can download the current stable version from CRAN:
+
+
+```{r, eval = FALSE}
+install.packages("CliquePercolation")
+```
+
+## Community Guidelines
+
+If you have recommendations for improving **CliquePercolation**, want to report issues or problems with the package, or need support, you can either open an [issue](https://github.com/LangeJens/CliquePercolation/issues) on GitHub or contact me directly via e-mail (<lange.jens@outlook.com>). If you consider contributing to **CliquePercolation**, you can also submit a [pull request](https://github.com/LangeJens/CliquePercolation/pulls).
+
+## Illustrative Examples
+
+The following includes two examples -- one for an undirected, weighted network and one for an undirected, unweighted network -- to illustrate parts of the package with real data. For a more extensive description of (a) the clique percolation algorithms for undirected, weighted as well as undirected, unweighted networks and (b) the workflow of the package with multiple, fictitious examples, see the package vignette, e.g., by running `vignette("CliquePercolation")` after installing and loading (`library(CliquePercolation)`) the package. Furthermore, the documentation of each function covers various other standalone examples with fictitious and real data that can be used to verify their functionality. An overview of all functions and links to their documentations are available by running `?'CliquePercolation'`.
+
+### The Obama network
+
+The package includes a data set with 10 evaluative reactions toward Barack Obama from 5,914 participants (for details see `?Obama`). This data set can be used to estimate a Gaussian Graphical Model (i.e., a regularized partial correlation network), which is an undirected, weighted network. Subsequently, the **CliquePercolation** package allows analyzing the community structure of the network.
+
+First, we will load the data, estimate the network with the package **bootnet**, and plot the network with the package **qgraph**.
+
+```{r, echo = TRUE, warning = FALSE, message = FALSE}
+#load data
+data(Obama)
+
+#estimate network
[…]
