### 2021-05-19 b7ebc7f
Adapt paper to new functions
