+++ paper/paper.md
@@ -12,7 +12,7 @@ authors:
 affiliations:
  - name: University of Hamburg
    index: 1
-date: "2021-01-06"
+date: "2021-05-19"
 bibliography: paper.bib
 ---
 
@@ -60,7 +60,7 @@ W <- qgraph::qgraph(W,
 
 The clique percolation algorithm proceeds in two steps. First, it identifies $k$-cliques in the network, i.e., fully conntected subgraphs with $k$ nodes, when the geometric mean of their edge weights exceeds the Intensity threshold $I$. Second, communities are defined as sets of adjacent $k$-cliques, i.e., $k$-cliques that share $k - 1$ nodes, allowing some nodes to be shared by communities or to be isolated.
 
-The package **CliquePercolation** faciliates executing these steps. First, it helps identifying optimal values for $k$ and $I$. For very small networks (as in \autoref{fig:fig1}), the entropy of the community partition should be maximized (treating isolated nodes as a separate community).
+The package **CliquePercolation** facilitates executing these steps. First, it helps identifying optimal values for $k$ and $I$. For very small networks (as in \autoref{fig:fig1}), the entropy of the community partition should be maximized (treating isolated nodes as a separate community).
 
 \begin{equation}
 Entropy = -\sum_{i=1}^N p_i * \log_2 p_i
@@ -95,8 +95,7 @@ and the `cpPermuteEntropy` function runs the permutation test
 
 
 ```r
-set.seed(4186)
-thresholds.permute <- cpPermuteEntropy(W, cpThreshold.object = threshold)
+thresholds.permute <- cpPermuteEntropy(W, cpThreshold.object = threshold, seed = 4186)
 ```
 
 \normalsize
@@ -111,10 +110,37 @@ returning the combinations of $k$ and $I$ that are more surprising than chance
 
 
 ```r
-thresholds.permute$Extracted.Rows
-#>    k Intensity Number.of.Communities Number.of.Isolated.Nodes Entropy.Threshold
-#> 22 3      0.09                     2                        1          1.418564
-#> 44 4      0.09                     1                        4          1.000000
+thresholds.permute
+#> 
+#> Confidence intervals for entropy values of random permutations of original network
+#> 
+#> --------------------
+#> 
+#> User-specified Settings
+#> 
+#> n = 100 
+#> interval = 0.95 
+#> CFinder = FALSE 
+#> ncores = 2 
+#> seed = 4186 
+#> 
+#> 
+#> --------------------
+#> 
+#> Confidence intervals
+#> 
+#>  k 95% CI lower 95% CI upper
+#>  3        1.166        1.299
+#>  4        0.113        0.267
+#> 
+#> 
+#> --------------------
+#> 
+#> Extracted rows from cpThreshold object
+#> 
+#>  k Intensity Number.of.Communities Number.of.Isolated.Nodes Entropy.Threshold
+#>  3      0.09                     2                        1             1.419
+#>  4      0.09                     1                        4             1.000
 ```
 
 \normalsize
@@ -130,65 +156,60 @@ The highest entropy results for $k = 3$ and $I = 0.09$, which can be used to run
 
 ```r
 cp.k3I.09 <- cpAlgorithm(W, k = 3, method = "weighted", I = 0.09)
-```
 
-\normalsize
-
-&nbsp;
-
-The algorithm identified two communities
-
-&nbsp;
-
-\small
-
-
-```r
-cp.k3I.09$list.of.communities.labels
-#> [[1]]
-#> [1] "d" "e" "f" "g"
+cp.k3I.09
 #> 
-#> [[2]]
-#> [1] "a" "b" "c" "d"
+#> Results of clique percolation community detection algorithm
+#> 
[…]
