### 2021-05-19 #3210 @LangeJens
@donaldRwilliams I addressed all issues in the repository. I again want to thank you for pushing me to implement these changes. As many of these things were entirely new to me, I hope you are nevertheless satisfied with how I implemented them. I also edited the vignette to cover the new functionality.

@karthik  I also adapted the paper to the new changes.
