### 2021-05-19 #3210 @donaldRwilliams
@LangeJens Wow !

I just took a look at the README and it is very (very) impressive.

I do like how all the changes look. Also note I suggested them in an effort to increase the impact of the package. e.g., before it was not clear to me how I might use the package. But now it is perfectly clear. I like it all a lot :-)
