### 2021-05-19 #3210 @LangeJens
@donaldRwilliams Thank you very much for your kind words! I agree that the changes you proposed will make the package more useful, accessible, and fitting to the classic approach in R. I always thought that the vignette is most helpful (and for some people it was), but I guess many people won't know about how to find it.
