### 2021-05-26 9b83634
Add package loading message
