+++ vignettes/CliquePercolation.Rmd
@@ -182,7 +182,7 @@ $$
 
 The logic behind $\chi$ is as follows. When $I$ is higher, there will be many equally small communities of size $k$. When $I$ is very low, there will be a gigantic community and many equally small ones. Thus, after excluding the largest community, for higher and smaller $I$, $\chi$ will be small. In contrast, when the community size distribution is broad (i.e., close to a power-law), then the variance of the communities (after excluding the largest community) will be higher. Thus, the point of the maximal variance is another way to optimize $I$ for different $k$.
 
-Nevertheless, also a stable estimate of $\chi$ requires a moderate number of communities. If the network is very small, such that only a few communities can be expected, also the $\chi$ threshold is unreliable. For instance, networks that represent psychological constructs such as attitudes [@Dalege+et+al:2016] or emotions [@Lange+et+al:2020], entail typically only up to 25 nodes.  Moreover, it would be undesirable to have, for instance, two communities of which one is twice as large as the other. Instead, equally sized communities are preferrable. An alternative way to optimize $I$ for different $k$ for very small networks would be to rely on the *entropy* of the community partition, for instance, an entropy measure based on Shannon Information [@Shannon:1948]. The idea is to find the most surprising community partition, defined as a low probability of knowing to which community a randomly picked node belongs. If there is only one community, surprisingness would be zero, because it would be certain that a randomly picked node belongs to this community. If there are, for instance, two communities of sizes 15 and 3, it would still not be very surprising, because most likely a randomly picked node will belong to the larger community. Surprisingness would be higher, however, when the communities are equal in size. Entropy based on Shannon Information captures this idea in the equation
+Nevertheless, also a stable estimate of $\chi$ requires a moderate number of communities. If the network is very small, such that only a few communities can be expected, also the $\chi$ threshold is unreliable. For instance, networks that represent psychological constructs such as attitudes [@Dalege+et+al:2016] or emotions [@Lange+et+al:2020], entail typically only up to 25 nodes.  Moreover, it would be undesirable to have, for instance, two communities of which one is twice as large as the other. Instead, equally sized communities are preferable. An alternative way to optimize $I$ for different $k$ for very small networks would be to rely on the *entropy* of the community partition, for instance, an entropy measure based on Shannon Information [@Shannon:1948]. The idea is to find the most surprising community partition, defined as a low probability of knowing to which community a randomly picked node belongs. If there is only one community, surprisingness would be zero, because it would be certain that a randomly picked node belongs to this community. If there are, for instance, two communities of sizes 15 and 3, it would still not be very surprising, because most likely a randomly picked node will belong to the larger community. Surprisingness would be higher, however, when the communities are equal in size. Entropy based on Shannon Information captures this idea in the equation
 
 $$
 Entropy = -\sum_{i=1}^N p_i * \log_2 p_i
@@ -420,7 +420,7 @@ The resulting object has the class `cpPermuteEntropy`. The printing function sho
 permute
 ```
 
-The resulting object is also a list with the two respective objects, assessible via
+The resulting object is also a list with the two respective objects, accessible via
 
 ```{r, echo = TRUE, results = FALSE}
 permute$Confidence.Interval
