### 2021-05-28 610c190
Set ncores in example to 1 and adapt date
