+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: CliquePercolation
 Version: 0.3.0
-Date: 2020-04-07
+Date: 2021-05-28
 Title: Clique Percolation for Networks
 Description: Clique percolation community detection for weighted and 
     unweighted networks as well as threshold and plotting functions.
