### 2021-05-29 82342e0
Fix DOIs and URLs
