+++ vignettes/bibtex.bib
@@ -34,7 +34,7 @@
   volume = {Complex Systems},
   pages = {1695},
   year = {2006},
-  url = {http://igraph.org},
+  url = {https://igraph.org},
 }
 
 @Article{Dalege+et+al:2016,
