### 2021-05-29 #3210 @LangeJens
@karthik I deposited the package in Zenodo. The doi is 10.5281/zenodo.4844318. I also submitted the package as version 0.3.0 to CRAN and they confirmed that it is soon online.
