### 2021-06-05 #3210 @Kevin-Mattheus-Moerman
@LangeJens I'll be helping to process acceptance and publication of this work. I have check the archive link and have proofread your paper. Below are some remaining issues that need your attention:

- [ ] Please ensure all Zenodo meta-data matches that of the JOSS paper, in particular ensure that the archive title matches the paper title. You should be able to edit this information on Zenodo.
- [ ] Please amend your affiliation to include city and country information.
- [ ] Check your first sentence `... one – if not the most – versatile field of research...`, it should be readable without `– if not the most –`, i.e. perhaps it should read: `... one of the  – if not the  – most versatile field of research...`
