+++ paper/paper.md
@@ -10,9 +10,9 @@ authors:
     orcid: 0000-0002-5375-3247
     affiliation: 1
 affiliations:
- - name: University of Hamburg
+ - name: University of Hamburg, Hamburg, Germany
    index: 1
-date: "2021-05-19"
+date: "2021-06-07"
 bibliography: paper.bib
 ---
 
@@ -20,13 +20,13 @@ bibliography: paper.bib
 
 # Summary and Statement of Need
 
-Modeling complex phenomena as networks constitutes one -- if not the most -- versatile field of research [@Barabasi:2011]. Indeed, many interconnected entities can be represented as networks, in which entities are called *nodes* and their connections are called *edges*. For instance, networks can represent friendships between people, hyperlinks between web pages, or correlations between questionnaire items. One structural characteristic of networks that is investigated frequently across various sciences is the detection of *communities* [@Fortunato:2010]. Communities are strongly connected subgraphs in the network such as groups of friends, thematic fields, or latent factors. Most community detection algorithms thereby put each node in only one community. However, nodes are often shared by multiple communities, e.g., when a person is part of multiple groups of friends, web pages belong to different thematic fields, or items load on multiple factors. The most popular community detection algorithm that is aimed at identifying such overlapping communities is the *clique percolation algorithm* [@Farkas+et+al:2007; @Palla+et+al:2005].
+Modeling complex phenomena as networks constitutes one of the -- if not the most -- versatile field of research [@Barabasi:2011]. Indeed, many interconnected entities can be represented as networks, in which entities are called *nodes* and their connections are called *edges*. For instance, networks can represent friendships between people, hyperlinks between web pages, or correlations between questionnaire items. One structural characteristic of networks that is investigated frequently across various sciences is the detection of *communities* [@Fortunato:2010]. Communities are strongly connected subgraphs in the network such as groups of friends, thematic fields, or latent factors. Most community detection algorithms thereby put each node in only one community. However, nodes are often shared by multiple communities, e.g., when a person is part of multiple groups of friends, web pages belong to different thematic fields, or items load on multiple factors. The most popular community detection algorithm that is aimed at identifying such overlapping communities is the *clique percolation algorithm* [@Farkas+et+al:2007; @Palla+et+al:2005].
 
-The clique percolation algorithm is not yet implemented in a package in **R** [@R:2020]. So far, the primary software for running the algorithm is the standalone program [**CFinder**](http://www.cfinder.org/), written in **C++** and **Java** [@Adamcsek+et+al:2006]. However, **CFinder** cannot be used to construct networks from data or to visualize the solutions of the algorithm, requiring the simultaneous use of other software such as **R**. Handling multiple programs impedes a smooth workflow. Next to **CFinder**, an **R** function for running one variant of the clique percolation algorithm is available in a [GitHub repository](https://github.com/angelosalatino/CliquePercolationMethod-R). However, it is not implemented in a package and lacks functions for optimizing parameters of the algorithm as well as plotting its results. **CliquePercolation** overcomes these limitations as it entails functions for helping to optimize parameters of the algorithm, running the algorithm, and plotting the results.
[…]
