### 2021-06-07 #3210 @arfon
This paper is an extra-special one for JOSS as it's the last paper @karthik is editing for JOSS. Thanks so much for all of your help over the last few years Karthik – and the [close to fifty papers](https://joss.theoj.org/papers/edited_by/@karthik) you've edited for us 😻:zap::rocket::boom:
